refactor: remove hook to download kubeconfig for secret
